feat: port to SubVerso
The compiler interaction and source highlighting features of Verso are
now extracted to a library that can be used with many different Lean
versions. It provides tools for writing Lean projects that use
multiple toolchains and including examples from them in the same text.

This provides an alternative to inline code blocks for more
complicated cases.