feat: deduplication of hover info (#115)
Replace the CSS-based hover solution with a JS-based one. Comapred to the CSS solution, it's better on touch, a better UI overall, and more difficult to maintain. But most importantly, it significantly reduces the size of rendered HTML - the Functional Induction blog post went from 1.88 M to 587 K bytes, and the in-progress reference manual becomes reasonable to serve remotely.