fix: incremental message logs - off-by-one fix and footgun removal
Fixes incremental reporting, which previously suffered from two
problems:

1. Snapshot message logs were restored only when states were to be
re-used, but message logs capture the log at the _start_ of the
elaboration step and thus should be unconditionally restored. This led
to the step immediately prior to buffer damage losing its messages.

2. It was inconvenient to disable incrementality for "finishing"
commands, so it didn't happen in practice. Now, the finishing command
is provided as a parameter to incrementallyElabCommand, so the
framework can arrange for it to be run correctly and the details of
promises that need resolving are not exposed. This led to the final
messages getting lost, while now they're saved. Footgun removal!