feat: improve legibility (#301)
* feat: reduce font size on desktop

Since we use `rem` almost everywhere, the effect on the whole page is
that it is slightly zoomed out. It doesn't increase the number of
characters per line.

16px is taken from MDN, the rust book, and stack overflow. It is the
most common font size on the web.

* feat: increase line-height for body content

I investigated line heights from MDN and the rust book, and picked the
lesser one out of caution. To me, it increases legibility.