fix: don't hide the header bar (#318)
These changes make it so that the title doesn't move from the header bar
to the table of contents by default. This means that some books will
have an overflowing header on small devices by default, but it will be
easy to notice and it is easy to fix.

The header bar needs to always be displayed, because that's where the
menu button lives.