fix: several small UI fixes (#338)
* fix: display the label consistently over namedocs headers

* fix: overflowing code examples, LawfulMonad

I think this is only displayed in a few places, one of them being
`LawfulMonad.{u, v}`. Would it be possible to apply some formatting to
those examples too, so the scrolling becomes unnecessary?

* fix: darker warning color (and centralize it)

I picked #E7A71D, as it has 2.11 contrast to white, which should be OK
for visual cues. The previous color was #efd871, which only has 1.42.
See
https://webaim.org/resources/contrastchecker/?fcolor=EFD871&bcolor=FFFFFF.