feat: redesign of definition boxes (#342)
Redesign the definition boxes so they look more like the example boxes.
This creates a more unified sense of style across the manuals.

This also introduces variables for the padding and vertical margin of
definition boxes and example boxes, to align them and make it more
apparent that they are connected.