fix: elaborate blocks as top-level defs
This causes each Verso block to be elaborated to a top-level
definition, which makes the checking for `sorry` and errors work the
way it should. Previously, the `sorry` warning was emitted in the
presence of errors, which was inconsistent with the rest of Lean.

This is also a step towards elaborating Verso directly, which will
probably improve performace in the future.