feat: alternative approach to incrementality (#400)
Replaces the Lean command parser with a new one that takes a block at
a time, rather than parsing Verso as one giant command and
implementing incrementality "by hand".

Each Verso block is elaborated to a top-level
definition, which makes the checking for `sorry` and errors work the
way it should. Without this, the `sorry` warning was emitted in the
presence of errors, which was inconsistent with the rest of Lean.

This is also a step towards elaborating Verso directly, which will
probably improve performance in the future.

A diff threshold was added to `leanOutput` in the manual genre to account for differences in IR shown in examples in the reference manual as a result of compiling blocks to separate defs.