chore: bump to nightly-2025-06-23 and merge `main` (#452)
Contained:

* feat: show hovers on closing delimiters (#428)

Closes #427

* feat: linter for Unicode typography (#430)

* chore: adapt to upstream SubVerso improvements (#432)

In particular, this fixes highlighting for sorts and improves the
display of tactic states.

* feat: LinkTargets for Blog Genre and definitions (#431)

* fix: parse ` ` correctly (#433)

* doc: update README for nightly procedures (#424)

* fix: hover for imax had flipped args (#434)

* fix: remove obsolete z-index customization for tactic states (#435)

* chore: bump SubVerso version (#438)

* fix: demo textbook would delete files (#439)

It seems that a relative path became absolute. Now it's relative again.

* chore: bump SubVerso for bug fixes (#440)

* Fix port number for accessing demosite in README.md (#442)

* feat: allow parts to disable showing inline HTML ToC (#444)

* feat: cross-references between documents (#441)

This is a feature inspired by InterSphinx, to allow documents to reference one another.

The idea is that the internal cross-references database is placed at a well-known URL, and documents can download and cache it, using it to resolve links.

* feat: add linters for extra markup (#448)

In particular, this will help users not get caught on the fact that
`*` is different from Markdown.

* fix: display multiple authors correctly (#449)

* feat: add part generation and info-string checking in Markdown elab (#445)

This PR adds the Markdown elaboration features required for error explanations, namely the ability to create `Part`s from Markdown headers, the ability to access infostrings in code-block elaborators, and a "strict" inline code elaboration mode.

* feat: linter for missing section tags in manual genre (#450)

* fix: respect `none` for section tags in manual genre (#451)

* chore: bump to nightly-2025-06-23

---------

Co-authored-by: Pim Otte <otte.pim@gmail.com>
Co-authored-by: Hanson Char <hanson.char@gmail.com>
Co-authored-by: jrr6 <7482866+jrr6@users.noreply.github.com>