feat: allow customising blog header (#472)
Currently, any page generated by the blog genre starts with the magic string `<!DOCTYPE html>`. This PR modifies that to, by default, use `Html.doctype` (as used by the manual genre). 
It also allows customising that string to something else, which is useful to replace it with a Jekyll-friendly front-matter. The default behaviour is left unchanged.