chore: bump to nightly-2025-07-27 (#481)
* feat: structured messages (#479)

Lean's messages are now showed as structured information, rather than strings, which enables hovers and interactive traces.

* fix: mobile Safari rendering issue (#478)

* fix: parser errors getting silently lost (#476)

* feat: allow customising blog header (#472)

Currently, any page generated by the blog genre starts with the magic string `<!DOCTYPE html>`. This PR modifies that to, by default, use `Html.doctype` (as used by the manual genre). 
It also allows customising that string to something else, which is useful to replace it with a Jekyll-friendly front-matter. The default behaviour is left unchanged.

* chore: bump to latest nightly

---------

Co-authored-by: Bhavik Mehta <bhavikmehta8@gmail.com>