fix: correct markdown header processing (#539)
Previously, parts generated by `addPartFromMarkdown` were at wrong nesting depth for nontrivial structures. Error explanations (for which this code was originally written) did not exercise this but because they have fairly simple structure, without many children of the same header, or decreases of header depth.

This PR simplifies the verso-nesting-depth-to-markdown-nesting-depth state, to a list of `Nat` rather than a list of pair of `Nat`s. I think without sacrificing correctness.