refactor: prepare to serialize highlights (#550)
Pulls out common code from the two inline generating functions and the two block generating functions, towards the goal of minimizing cut and paste when we switch from just $(quote hlt) to a more complicated serialization procedure for including highlighted material in files.

Eliminates some dead code.

Fixes a small bug that meant that -show didn't work on lean or inst roles in VersoManual. This is broken on the current release candidate, but I don't think the code path was ever exercised and noone noticed