refactor: reduce output C size with redundancy elimination and serialization (#552)
`Highlighted` objects contain highly redundant information, and the
result is that the C code we output at present redundantly represents
this redundant information.

This PR steals a trick that @david-christiansen introduced in
teorth/analysis#352 of exporting `Highlighted` to the existing
`ExportCode` format, serializing the resulting data to JSON, and
producing syntax that transforms that resulting string back to
Highlighted. (Aggressive compiler optimizations could undo this work,
but that does not seem to be an issue at present.)

Leads to a 11% reduction in the output size of C code for the reference
manual (as reported by `find .lake/build/ir -name "*.c" | xargs du -c |
tail -n 1`) — 1049888 to 936912.

---------

Co-authored-by: David Thrane Christiansen <david@lean-fro.org>