feat: JS source maps
This PR adds JavaScript source maps to JavaScript files included in
Verso output. This fixes issues with browsers looking for them and not
finding them.

In the process, rather than fixing lots of copy-pasted JS specs, a
"mixin" mechanism was added to the block_extension and
inline_extension commands in the manual genre that allows functions to
be used to rewrite block and inline specs.