fix: relativize paths before navigating to them (#578)
I'd misremembered/interpreted/understood the [URL Web
API](https://developer.mozilla.org/en-US/docs/Web/API/URL/URL) and the
[<base> tag
spec](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/base)
— if you give an absolute path (starting with a forward slash) then all
these will navigate relative to the website root — the full base URL in
both cases only matters for relative paths.

Because VersoManual's find page **is** setting the website base to be
the "root-of-the-manual", we can just avoid constructing the full URL
and navigate to the appropriate relative path, but first we have to
strip the slash (that we probably shouldn't be storing in the first
place) from the search result.

Also adjusts the leanfile to somewhat-adequately capture the actual
text-file dependencies of Verso.