refactor: create an intermediate 'VersoDoc' data structure (#594)
A `VersoDoc genre` is intended to represent an *unevaluated* `Part
genre`. The result of a `#doc` command stops being syntax denoting a
value of type `Part genre`, and becomes syntax denoting a value of type
`VersoDoc genre`.

Presenting this indirection layer allows us for more flexibility in
serializing some or all of a document in a way that minimally affects
existing code. There are two alternative designs here that I considered:

1) Having `VersoDoc.toPart` be monadic (returning `m (Part genre)`
instead of `Part genre`), since we might in the future want to have a
part that includes a reference to an auxiliary file or data structure.
Doing this seemed like it would require much more substantial changes to
existing code.
2) Having `VersoDoc.toPart` require an extra argument, a piece of
context that can be threaded through that might represent something like
serialization data in the future. This would be somewhat simpler to
implement (it would be implementing the reader monad instead of an
arbitrary monad).