refactor: reduce emitted C code sizes for manual genre (#609)
Adds three things that work together:

- A `SubVerso.Highlighting.Exporting` argument to `DocElabM.State`,
which evaluates every Lean `Highlighting` block in the context of a
growing deduplication table. This is an option type, if it's set to the
default `.none` (in any non-manual genre), inline lean will continue
using per-code-fence redundancy elimination.
- A `SubVerso.Highlighting.Export` field for `DocReconstruction`, which
can be referenced during doc reconstruction to reconstruct a
`Highlighting` object from a single table key
- A `String` field for `VersoDoc.mk` that is intended to contain a Json
object. When `VersoDoc.toPart` runs, it looks for a `"highlight"` key in
that object, and it is is present uses it to reconstruct the
`SubVerso.Highlighting.Export` object that is then used for document
reconstruction.

This is doing something very _specific_, and having this functionality —
building up a custom type during elaboration, serializing it, and then
making it available during document reconstruction — could easily be
applicable to other use cases and genres. I think we should absolutely
add the more generic functionality as soon as we have a second use case
that uses it.

---------

Co-authored-by: David Thrane Christiansen <david@lean-fro.org>