chore: refactor HTML asset handling in config and state for Manual
This factors out common components of the manual genre's handling of
HTML assets such as JS and CSS. Before, there were a series of
repeated fields; now, the fields are in a single structure. This
reduces the risk of errors due to forgetting to initialize one of
them, reduces unnecessary type conversions (sometimes they would be an
array, sometimes a set), and makes the code easier to follow.

Other repeated code, such as emitting these assets to disk, was also
factored out into a single function.

In the process, the handling of features such as the search box and
KaTeX were moved from functions that add code to the config into flags
in the HTML setup. This means that users who specify a config other
than the default won't suddenly lose math support if they don't
recapitulate the call to addKaTeX.