feat: implement basic syntax highlighting (#626)
For now, have a separate function for converting `Highlighted` into TeX
that can rely on it being used in the context of a `\Verbatim`
environment, meaning that it doesn't need to wrap every token in, e.g.,
something like `\texttt`. Possibly we may want to move everything to
using this in-verbatim-context function, but I haven't committed to it
yet.

Use standard TeX logger

---------

Co-authored-by: David Thrane Christiansen <david@davidchristiansen.dk>